pyproject.toml
[tool.poetry]
name = "fiveaxis"
version = "0.1.0"
description = ""
authors = ["Gopi Dutt V"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "^9.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
